Moodified code for CRUD fuctionality

+++ Shree_API_AWS/DataTransferObjects/Employee_DTO.cs
@@ -16,7 +16,7 @@
 
         public string? Ecname { get; set; }
 
-        public int? Ecnumber { get; set; }
+        public long? Ecnumber { get; set; }
 
         public string? Position { get; set; }
